requirements.txt
annotated-types==0.6.0
anyio==4.0.0
asgiref==3.7.2
attrs==23.1.0
blinker==1.7.0
boto3==1.28.78
botocore==1.31.78
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
convcolors==2.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.5
deprecation==2.1.0
Django==4.2.7
django-cors-headers==4.3.0
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
dynaconf==3.2.4
entrypoints==0.4
exceptiongroup==1.1.3
extcolors==1.0.0
flatbuffers==23.5.26
gitdb==4.0.11
GitPython==3.1.40
gotrue==1.3.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
humanfriendly==10.0
hvac==2.0.0
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
itypes==1.2.0
jmespath==1.0.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kaleido==0.2.1
knox==0.1.14
lazy_loader==0.3
llvmlite==0.40.1
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
openapi-codec==1.3.2
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==1.5.3
Pillow==9.3.0
platformdirs==3.11.0
plotly==5.18.0
pooch==1.8.0
postgrest==0.13.0
protobuf==3.20.3
pyarrow==14.0.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyMatting==1.1.10
Pympler==1.0.1
pypng==0.20220715.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
realtime==1.0.0
referencing==0.30.2
rembg==2.0.50
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
s3transfer==0.7.0
scikit-image==0.22.0
scipy==1.11.3
shortuuid==1.0.11
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sqlparse==0.4.4
storage3==0.6.1
StrEnum==0.4.15
supabase==2.0.3
supafunc==0.3.1
sympy==1.12
tenacity==8.2.3
tifffile==2023.9.26
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.18
validators==0.22.0
watchdog==3.0.0
websockets==10.4
whitenoise==6.6.0
win32-setctime==1.1.0
zipp==3.17.0